requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
python-dotenv
urllib3
dspy-ai>=2.5.0
pydantic>=2.0.0